*: cleanup and adjust the directory struct
go code is move to cmd/cora
remove the files that are definitely not used
