+++ cmd/cora/dump.go
@@ -1,4 +1,4 @@
-package cora
+package main
 
 import (
 	"errors"
