+++ cmd/cora/eval_test.go
@@ -1,4 +1,4 @@
-package cora
+package main
 
 import (
 	"os"
