--- go.mod
@@ -1,3 +1,0 @@
-module github.com/tiancaiamao/cora
-
-go 1.19
